[WIP] almost everything runs fine
