+++ lib/cli/cmd/compile.js
@@ -9,6 +9,7 @@
 
   var path = require('path');
   var async = require('async');
+  var mkdirp = require('mkdirp');
 
   var compiler = require('../../parsers/compiler');
   var utils = require('../utils');
@@ -42,12 +43,37 @@
     }
   };
 
+  var fetchContent = function(data, callback) {
+    var files = [];
+    utils.walkDir(data.sourceDir, {}, function(sourcePath, done) {
+      fs.readFile(sourcePath, function(err, content) {
+        if (err) {
+          return done(err);
+        }
+        
+        files.push({
+          name: sourcePath,
+          contents: content.toString()
+        });
+        done();
+      })
+    },function(err) {
+      /* istanbul ignore if */
+      if (err) {
+        return callback(err);
+      } else {
+        data.files = files;
+        callback(null, data);
+      }
+    })
+  }
+
   var warnAndCompileGame = function(data, callback) {
     if (data.needsCompilation) {
       if (data.command !== 'compile') {
         console.log('Game file is out of date, recompiling.'.red);
       }
-      compiler.compileGame(data.sourceDir, function(err, game) {
+      compiler.compileGame(data.files, function(err, game) {
         if (err) {
           return callback(err);
         }
@@ -67,7 +93,31 @@
       return callback(null, data);
     }
     var indent = data.indent ? 2 : 0;
-    compiler.saveCompiledGame(
+
+    var _saveCompiledGame = function(game, outPath, indent, callback) {
+      convertGameToJSON(game, indent, function(err, json) {
+        /* istanbul ignore if */
+        if (err) {
+          return callback(err);
+        }
+        var diry = path.dirname(outPath);
+        mkdirp(diry, function(err) {
+          /* istanbul ignore if */
+          if (err) {
+            return callback(err);
+          }
+          fs.writeFile(outPath, json, function(err) {
+            /* istanbul ignore if */
+            if (err) {
+              return callback(err);
+            }
+            callback(null);
+          });
+        });
+      });
+    };
+
+    _saveCompiledGame(
       data.game, data.compiledPath, indent, function(err) {
         if (err) {
           return callback(err);
@@ -77,6 +127,8 @@
       }
     );
   };
+
+  
 
   // ----------------------------------------------------------------------
   // Compile: Checks a project is correct and compiles.
@@ -115,7 +167,9 @@
     };
     async.waterfall([getData, ensureProject,
                      checkCompiledGameUpToDate,
+                     fetchContent,
                      warnAndCompileGame,
+                     convertGameToJSON,
                      saveCompiledGame],
                     callback);
   };
