+++ lib/cli/cmd/make-book.js
@@ -13,14 +13,13 @@
   var async = require('async');
 
   var utils = require('../utils');
-  var compiler = require('../../parsers/compiler');
   var cmdCompile = require('./compile').cmd;
   var gamebook = require('../../search/gamebook');
   var html = require('../../ui/content/html');
   var latex = require('../../ui/content/latex');
 
   var loadGameAndSource = function(data, callback) {
-    compiler.loadCompiledGame(data.compiledPath, function(err, game, json) {
+    utils.loadCompiledGame(data.compiledPath, function(err, game, json) {
       if (err) {
         return callback(err);
       }
