+++ lib/cli/cmd/make-html.js
@@ -14,11 +14,10 @@
   var uglify = require('uglify-js');
 
   var utils = require('../utils');
-  var compiler = require('../../parsers/compiler');
   var cmdCompile = require('./compile').cmd;
 
   var loadGameAndSource = function(data, callback) {
-    compiler.loadCompiledGame(data.compiledPath, function(err, game, json) {
+    utils.loadCompiledGame(data.compiledPath, function(err, game, json) {
       if (err) {
         return callback(err);
       }
