+++ lib/cli/cmd/random-test.js
@@ -15,13 +15,12 @@
   var cmdCompile = require('./compile').cmd;
   var CLUserInterface = require('../../ui/cli').CommandLineUserInterface;
   var RandomTestPrompt = require('../../ui/random_prompt').RandomTestPrompt;
-  var compiler = require('../../parsers/compiler');
 
   var loadGame = function(data, callback) {
     if (data.needsCompilation) {
       return callback(null, data);
     }
-    compiler.loadCompiledGame(data.compiledPath, function(err, game) {
+    utils.loadCompiledGame(data.compiledPath, function(err, game) {
       if (err) {
         return callback(err);
       }
