+++ lib/cli/cmd/run.js
@@ -13,13 +13,12 @@
   var utils = require('../utils');
   var cmdCompile = require('./compile').cmd;
   var CLUserInterface = require('../../ui/cli').CommandLineUserInterface;
-  var compiler = require('../../parsers/compiler');
 
   var loadGame = function(data, callback) {
     if (data.needsCompilation) {
       return callback(null, data);
     }
-    compiler.loadCompiledGame(data.compiledPath, function(err, game) {
+    utils.loadCompiledGame(data.compiledPath, function(err, game) {
       if (err) {
         return callback(err);
       }
