+++ lib/cli/utils.js
@@ -9,6 +9,7 @@
 
   var path = require('path');
   var fs = require('fs');
+  var dive = require('dive');
   var handlebars = require('handlebars');
   var _ = require('lodash');
 
@@ -58,7 +59,7 @@
     pattern = pattern || /\.dry$/;
 
     var latest = null;
-    compiler.walkDir(diry, {}, function(srcpath, done) {
+    walkDir(diry, {}, function(srcpath, done) {
       // Skip non relevant files.
       if (!pattern.test(srcpath)) {
         return done();
@@ -166,13 +167,13 @@
       if (err) {
         return callback(err);
       }
-      compiler.walkDir(
+      walkDir(
         sourceDir, {directories:true, files:false}, mkDir,
         function(err) {
           if (err) {
             return callback(err);
           }
-          compiler.walkDir(sourceDir, {}, copyFile, function(err) {
+          walkDir(sourceDir, {}, copyFile, function(err) {
             if (err) {
               return callback(err);
             } else {
@@ -186,6 +187,83 @@
     data.year = (new Date()).getFullYear();
     mkDir(sourceDir, doCopy);
   };
+
+  /* An augmented version of dive that allows the processing 'action'
+   * function to be asynchronous. It should accept a parameter
+   * 'finished' which should be called when it has completed its work,
+   * with an error if something happened. Errors are not passed along
+   * to future actions, but actions are stopped and the finish routine
+   * is called. Unlike dive, the finished routine is guaranteed to be
+   * called after all actions have completed, and is passed the error,
+   * if there is one.
+   *
+   * action(src, done); done(err)
+   * finished(err)
+   */
+  var walkDir = function(diry, opts, action, finished) {
+    var hasFinishedDive = false;
+    var donesPending = 0;
+    var error;
+    var completed = false;
+    var complete = function() {
+      if (!completed) {
+        completed = true;
+        if (error !== undefined) {
+          return finished(error);
+        } else {
+          return finished();
+        }
+      }
+    };
+    dive(diry, opts, function(err, srcpath) {
+      if (err) {
+        error = err;
+        /* istanbul ignore else */
+        if (donesPending === 0) {
+          complete();
+        }
+      } else if (!error) {
+        donesPending++;
+        action(srcpath, function(err) {
+          if (err) {
+            error = err;
+          }
+          donesPending--;
+          // This may be untestable, because of the async nature of the walk.
+          /* istanbul ignore next */
+          if (donesPending === 0 && hasFinishedDive) {
+            complete();
+          }
+        });
+      }
+    }, function() {
+      // Only called when no error is passed to the dive action callback.
+      hasFinishedDive = true;
+      // This may be untestable, because of the async nature of the walk.
+      /* istanbul ignore next */
+      if (donesPending === 0) {
+        complete();
+      }
+    });
+  };
+
+  var loadCompiledGame = function(inPath, callback) {
+    fs.readFile(inPath, function(err, json) {
+      /* istanbul ignore if */
+      if (err) {
+        return callback(err);
+      }
+      compiler.convertJSONToGame(json, function(err, game) {
+        /* istanbul ignore if */
+        if (err) {
+          return callback(err);
+        } else {
+          return callback(null, game, json);
+        }
+      });
+    });
+  };
+
 
   // ----------------------------------------------------------------------
 
@@ -213,6 +291,8 @@
     getLatestMTime: getLatestMTime,
     isUpToDate: isUpToDate,
     copyTemplate: copyTemplate,
+    loadCompiledGame: loadCompiledGame,
+    walkDir: walkDir,
 
     Command: Command
   };
