+++ lib/parsers/compiler.js
@@ -10,10 +10,7 @@
   var assert = require('assert');
   var _ = require('lodash');
   var async = require('async');
-  var dive = require('dive');
   var fs = require('fs');
-  var path = require('path');
-  var mkdirp = require('mkdirp');
 
   var infoParser = require('./info');
   var sceneParser = require('./scene');
@@ -548,118 +545,10 @@
 
   // ----------------------------------------------------------------------
 
-  /* An augmented version of dive that allows the processing 'action'
-   * function to be asynchronous. It should accept a parameter
-   * 'finished' which should be called when it has completed its work,
-   * with an error if something happened. Errors are not passed along
-   * to future actions, but actions are stopped and the finish routine
-   * is called. Unlike dive, the finished routine is guaranteed to be
-   * called after all actions have completed, and is passed the error,
-   * if there is one.
-   *
-   * action(src, done); done(err)
-   * finished(err)
-   */
-  var walkDir = function(diry, opts, action, finished) {
-    var hasFinishedDive = false;
-    var donesPending = 0;
-    var error;
-    var completed = false;
-    var complete = function() {
-      if (!completed) {
-        completed = true;
-        if (error !== undefined) {
-          return finished(error);
-        } else {
-          return finished();
-        }
-      }
-    };
-    dive(diry, opts, function(err, srcpath) {
-      if (err) {
-        error = err;
-        /* istanbul ignore else */
-        if (donesPending === 0) {
-          complete();
-        }
-      } else if (!error) {
-        donesPending++;
-        action(srcpath, function(err) {
-          if (err) {
-            error = err;
-          }
-          donesPending--;
-          // This may be untestable, because of the async nature of the walk.
-          /* istanbul ignore next */
-          if (donesPending === 0 && hasFinishedDive) {
-            complete();
-          }
-        });
-      }
-    }, function() {
-      // Only called when no error is passed to the dive action callback.
-      hasFinishedDive = true;
-      // This may be untestable, because of the async nature of the walk.
-      /* istanbul ignore next */
-      if (donesPending === 0) {
-        complete();
-      }
-    });
-  };
 
-  // data is a dict of {filename : string containing file data}
-  // TODO: implement this
-  var compileGameFromData = function(filenames, dryData, infoData, callback) {
-    var info = infoParser.parseFromContent(infoData.fileName, infoData.data, function(err, dry) {
-    });
-    // prevScene is a special scene that is always accessible and blank.
-    // TODO: have a better way of displaying special scenes?
-    var prevScene = {id: 'prevScene',
-                     content: {content: '', type: 'paragraph'}};
-    var prevTopScene = {id: 'prevTopScene',
-                     content: {content: '', type: 'paragraph'}};
-    var jumpScene = {id: 'jumpScene',
-                     content: {content: '', type: 'paragraph'}};
-    // returnScene is used with goSub...
-    var returnScene = {id: 'returnScene',
-                       content: {content: '', type: 'paragraph'}};
-    // backSpecialScene should only be used within special scenes,
-    // and is used to go back to the previous non-special scene.
-    var backSpecialScene = {id: 'backSpecialScene',
-                            content: {content: '', type: 'paragraph'}};
-    var scenes = [prevScene, prevTopScene, jumpScene, backSpecialScene, returnScene];
-    var qualities = [];
-    var qdisplays = [];
-    var callbackFn = function(err, result) {
-          /* istanbul ignore if */
-          if (err) {
-            return done(err);
-          }
-          info = result;
-          done();
-    };
-    var dryCallback = function(err, result) {
-          /* istanbul ignore if */
-          if (err) {
-            return done(err);
-          }
-          scenes.push(result);
-          done();
-    };
-    for (var i = 0; i < filenames.length; i++) {
-        var fn = filenames[i];
-        if (/(^|\/)info\.dry$/.test(fn)) {
-            infoParser.parseFromContent(fn, dryData[fn], callbackFn);
-        } else if (/\.scene\.dry$/.test(fn)) {
-            dryParser.parseFromContent(fn, dryData[fn], dryCallback);
-        }
-    }
-  };
-
-  // Compiles a directory of dry data files into a game object.
-  // TODO: refactor this to use a list of file objects rather than html?
-  // just a list of strings...
-  var compileGame = function(sourceDir, callback) {
+  // Compiles a list of dry data files into a game object.
+  // A data file is a dict of the form {name : "filename.type.dry", contents: "string containing file data"}
+  var compileGame = function(dry_files_contents, callback) {
     var info;
     // prevScene is a special scene that is always accessible and blank.
     // TODO: have a better way of displaying special scenes?
@@ -679,101 +568,54 @@
     var scenes = [prevScene, prevTopScene, jumpScene, backSpecialScene, returnScene];
     var qualities = [];
     var qdisplays = [];
-    walkDir(sourceDir, {}, function(sourcePath, done) {
+    for(const file of dry_files_contents) {
+      const {name:sourcePath, contents} = file;
+
       if (/(^|\/)info\.dry$/.test(sourcePath)) {
-        infoParser.parseFromFile(sourcePath, function(err, result) {
+        infoParser.parseFromContent(sourcePath, contents, function(err, result) {
           /* istanbul ignore if */
           if (err) {
-            return done(err);
+            return callback(err);
           }
           info = result;
-          done();
         });
       } else if (/\.scene\.dry$/.test(sourcePath)) {
-        sceneParser.parseFromFile(sourcePath, function(err, result) {
+        sceneParser.parseFromContent(sourcePath, contents, function(err, result) {
           /* istanbul ignore if */
           if (err) {
-            return done(err);
+            return callback(err);
           }
           scenes.push(result);
-          done();
         });
       } else if (/\.quality\.dry$/.test(sourcePath)) {
-        qualityParser.parseFromFile(sourcePath, function(err, result) {
+        qualityParser.parseFromContent(sourcePath, contents, function(err, result) {
           /* istanbul ignore if */
           if (err) {
-            return done(err);
+            return callback(err);
           }
           qualities.push(result);
-          done();
         });
       } else if (/\.qdisplay\.dry$/.test(sourcePath)) {
-        qdisplayParser.parseFromFile(sourcePath, function(err, result) {
+        qdisplayParser.parseFromContent(sourcePath, contents, function(err, result) {
           /* istanbul ignore if */
           if (err) {
-            return done(err);
+            return callback(err);
           }
           qdisplays.push(result);
-          done();
         });
       } else {
         // Skip this file.
-        done();
+        callback();
       }
-    }, function(err) {
-      /* istanbul ignore if */
-      if (err) {
-        return callback(err);
-      } else {
-        compile(info, scenes, qualities, qdisplays, function(err, game) {
-          /* istanbul ignore if */
-          if (err) {
-            return callback(err);
-          }
-          callback(null, game);
-        });
-      }
-    });
-  };
+    }
 
-  var loadCompiledGame = function(inPath, callback) {
-    fs.readFile(inPath, function(err, json) {
+    compile(info, scenes, qualities, qdisplays, function(err, game) {
       /* istanbul ignore if */
       if (err) {
         return callback(err);
       }
-      convertJSONToGame(json, function(err, game) {
-        /* istanbul ignore if */
-        if (err) {
-          return callback(err);
-        } else {
-          return callback(null, game, json);
-        }
-      });
-    });
-  };
-
-  var saveCompiledGame = function(game, outPath, indent, callback) {
-    convertGameToJSON(game, indent, function(err, json) {
-      /* istanbul ignore if */
-      if (err) {
-        return callback(err);
-      }
-      var diry = path.dirname(outPath);
-      mkdirp(diry, function(err) {
-        /* istanbul ignore if */
-        if (err) {
-          return callback(err);
-        }
-        fs.writeFile(outPath, json, function(err) {
-          /* istanbul ignore if */
-          if (err) {
-            return callback(err);
-          }
-          callback(null);
-        });
-      });
-    });
+      callback(null, game);
+    })
   };
 
   var convertGameToJSON = function(game, indent, callback) {
@@ -808,10 +650,7 @@
   module.exports = {
     compile: compile,
     getCandidateAbsoluteIds: getCandidateAbsoluteIds,
-    walkDir: walkDir,
     compileGame: compileGame,
-    loadCompiledGame: loadCompiledGame,
-    saveCompiledGame: saveCompiledGame,
     convertJSONToGame: convertJSONToGame,
     convertGameToJSON: convertGameToJSON
   };
