+++ lib/parsers/dry.js
@@ -8,7 +8,6 @@
   'use strict';
 
   var _ = require('lodash');
-  var path = require('path');
   var fs = require('fs');
   var assert = require('assert');
 
@@ -80,11 +79,9 @@
     }
   };
 
-  var filenameRe = /^([\w-]+?(\.[\w-]+?)*?)(\.([\w-]+?))?$/;
+  var filenameRe = /^([\w-]+?(\.[\w-]+?)*?)(\.([\w-]+?))?\.([^.]*)$/;
   var parseFilename = function(filename) {
-    var ext = path.extname(filename);
-    var base = path.basename(filename, ext);
-    var match = filenameRe.exec(base);
+    var match = filenameRe.exec(filename);
     if (match) {
       var obj = {
         id: match[1],
@@ -423,16 +420,6 @@
     return callback(null, result);
   };
 
-  // Loads and parses the given file.
-  var parseFromFile = function(filename, callback) {
-    fs.readFile(filename, function(err, content) {
-      if (err) {
-        return callback(err);
-      }
-      parseFromContent(filename, content.toString(), callback);
-    });
-  };
-
   // ======================================================================
 
   var removeMetadataFromArray = function(array) {
@@ -496,7 +483,6 @@
 
   module.exports = {
     parseFromContent: parseFromContent,
-    parseFromFile: parseFromFile,
 
     regexes: {
       id: idRe, idString: idReString,
