+++ lib/parsers/make.js
@@ -33,25 +33,12 @@
     };
   };
 
-  var makeParseFromFile = function(schema) {
-    var ensure = validators.makeEnsureObjectMatchesSchema(schema);
-    return function(filename, callback) {
-      dryParser.parseFromFile(filename, function(err, dry) {
-        if (err) {
-          return callback(err);
-        }
-        ensure(dry, null, callback);
-      });
-    };
-  };
-
   var makeParseFromDry = validators.makeEnsureObjectMatchesSchema;
 
   var makeExports = function(schema) {
     return {
       schema: schema,
       parseFromContent: makeParseFromContent(schema),
-      parseFromFile: makeParseFromFile(schema),
       parseFromDry: makeParseFromDry(schema)
     };
   };
@@ -60,7 +47,6 @@
     extendSchema: extendSchema,
 
     makeParseFromContent: makeParseFromContent,
-    makeParseFromFile: makeParseFromFile,
     makeParseFromDry: makeParseFromDry,
     makeExports: makeExports
   };
