+++ test/test_parsers_info.js
@@ -8,12 +8,24 @@
   'use strict';
 
   var path = require('path');
+  var fs = require('fs');
   var should = require('should');
   // Disable errors from using the should library.
   /*jshint -W030 */
 
   var parse = require('../lib/parsers/info');
   var dryParser = require('../lib/parsers/dry');
+
+
+  // Loads and parses the given file.
+  var parseFromFile = function(parseFromContent, filename, callback) {
+    fs.readFile(filename, function(err, content) {
+      if (err) {
+        return callback(err);
+      }
+      parseFromContent(filename, content.toString(), callback);
+    });
+  };
 
   describe('info parser', function() {
 
@@ -99,7 +111,7 @@
 
     it('should load and parse info file', function(done) {
       var fn = path.join(__dirname, 'files', 'test_info_parser.info.dry');
-      parse.parseFromFile(fn, function(err, result) {
+      parseFromFile(parse.parseFromContent, fn, function(err, result) {
         (!!err).should.be.false;
         dryParser.removeMetadataFromObject(result);
         result.should.eql({
@@ -114,7 +126,7 @@
 
     it('should pass on dry file errors', function(done) {
       var fn = path.join(__dirname, 'files', 'unknown.scene.dry');
-      parse.parseFromFile(fn, function(err, result) {
+      parseFromFile(parse.parseFromContent, fn, function(err, result) {
         (!!err).should.be.true;
         (result === undefined).should.be.true;
         done();
