+++ test/test_parsers_scene.js
@@ -8,6 +8,7 @@
   'use strict';
 
   var path = require('path');
+  var fs = require('fs');
   var should = require('should');
   // Disable errors from using the should library.
   /*jshint -W030 */
@@ -21,6 +22,16 @@
 
   var parse = require('../lib/parsers/scene');
   var dryParser = require('../lib/parsers/dry');
+
+  // Loads and parses the given file.
+  var parseFromFile = function(parseFromContent, filename, callback) {
+    fs.readFile(filename, function(err, content) {
+      if (err) {
+        return callback(err);
+      }
+      parseFromContent(filename, content.toString(), callback);
+    });
+  };
 
   describe('scene parser', function() {
 
@@ -215,7 +226,7 @@
 
     it('should load and parse scene file', function(done) {
       var fn = path.join(__dirname, 'files', 'test_scene_parser.scene.dry');
-      parse.parseFromFile(fn, function(err, result) {
+      parseFromFile(parse.parseFromContent, fn, function(err, result) {
         noerr(err);
         dryParser.removeMetadataFromObject(result);
         result.should.eql({
@@ -245,7 +256,7 @@
 
     it('should pass on dry file errors', function(done) {
       var fn = path.join(__dirname, 'files', 'unknown.scene.dry');
-      parse.parseFromFile(fn, function(err, result) {
+      parseFromFile(parse.parseFromContent, fn, function(err, result) {
         (!!err).should.be.true;
         (result === undefined).should.be.true;
         done();
